Add 'C++11: Syntax and Feature', a C++11 reference book for Japanese
'C++11: Syntax and Feature' is a C++11 reference book based on the
latest C++ specification, ISO/IEC 14882:2011.  It is written by Ezoe
Ryo, who is a member of C++ standard committee.  This book is fantastic
because the words used in this book is the same ones in C++
specification, so readers can look up them in specification easily.
It is published with GFDL license today.